Move image_coordinates_to_coordinates from utils to s2p
